+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.0
+### Added
+- Added SPM support
+
 ## 1.0.5
 
 ### Bug Fixes